Improvements
##Changes:
##-Changed the shown name in Steam of the games to be the name from steamgrid instead of the one in the executable
##-Added GUI
##-Inference of the steam_user_data_path so only 3 parameters are needed
##-Storage of the variables in a json file at the script location
##-Changed slightly the messages to be logged, now it informs if a titles is being skipped when it's already present in steam.
##-Added exceptions to the names of executables to be found to avoid using the wrong one in Unity games

##Limitations:
##Only 1 folder of games can be used. MultiDirFolder is bugged in Gooey and the fix hasn't been merged.
##In case it worked, the logic would need to be altered, as we use the path as is when joining paths
